Undefine _POSIX_C_SOURCE to fix socket.inet_aton()
socket.inet_aton() in Python 2.5.6 was returning 8-byte strings when
built with Pythonbrew on OS X 10.7.3.  The 8-byte strings were a result
of a broken type cast identified in:

- http://bugs.python.org/issue767150
- http://bugs.python.org/issue1008086

Part of the problem was that autoconf was failing to pick up on C lib's
inet_aton(), exposing the broken type cast mentioned in the bug reports
above.  I found an existing precedence for suppressing _POSIX_C_SOURCE
on older Darwins, which, when extended to cover 10.7.3, magically makes
this bug go away.